# chore: upgrade to typescript-eslint v8