--- a/packages/eslint-plugin/src/rules/ban-context-export/index.js
+++ b/packages/eslint-plugin/src/rules/ban-context-export/index.js
@@ -7,6 +7,8 @@ const { getTypeServices } = require('../../utils/type-services');
 
 /** @typedef { import('@typescript-eslint/utils').TSESTree.VariableDeclarator } VariableDeclarator*/
 /** @typedef { import('@typescript-eslint/utils').TSESTree.ExportSpecifier} ExportSpecifier */
+/** @typedef { import('@typescript-eslint/utils').TSESLint.RuleMetaDataDocs} RuleMetaDataDocs*/
+/** @typedef { import('@typescript-eslint/utils').TSESLint.RuleRecommendation} RuleRecommendation */
 /**
  * @typedef {{
  *  exclude?: string[];
@@ -33,10 +35,10 @@ module.exports = createRule({
       },
     ],
     type: 'problem',
-    docs: {
+    docs: /** @type {RuleMetaDataDocs & { recommended: RuleRecommendation }} */ ({
       description: 'Ban export of React context or context selector objects',
       recommended: 'recommended',
-    },
+    }),
     messages: {
       nativeContext: '{{exportName}} should not be exported directly',
       contextSelector: '{{exportName}} should not be exported directly',