--- a/packages/eslint-plugin/src/rules/ban-instanceof-html-element/index.js
+++ b/packages/eslint-plugin/src/rules/ban-instanceof-html-element/index.js
@@ -4,17 +4,18 @@ const createRule = require('../../utils/createRule');
 
 /**
  * @typedef {import('./types').HTMLElementConstructorName} HTMLElementConstructorName
- *
+ * @typedef { import('@typescript-eslint/utils').TSESLint.RuleMetaDataDocs} RuleMetaDataDocs
+ * @typedef { import('@typescript-eslint/utils').TSESLint.RuleRecommendation} RuleRecommendation
  */
 
 module.exports = createRule({
   name: 'ban-instanceof-html-element',
   meta: {
     type: 'problem',
-    docs: {
+    docs: /** @type {RuleMetaDataDocs & { recommended: RuleRecommendation }} */ ({
       description: 'Ban usage of instanceof HTMLElement comparison',
       recommended: 'recommended',
-    },
+    }),
     messages: {
       invalidBinaryExpression: 'instanceof {{right}} should be avoided, use isHTMLElement instead.',
     },