--- a/packages/fluentui/react-bindings/src/context-selector/useContextSelector.ts
+++ b/packages/fluentui/react-bindings/src/context-selector/useContextSelector.ts
@@ -53,7 +53,7 @@ export const useContextSelector = <Value, SelectedValue>(
         }
 
         return [payload[1], nextSelected] as const;
-      } catch (e) {
+      } catch {
         // ignored (stale props or some other reason)
       }
       return [...prevState] as const; // schedule update