--- a/tools/eslint-rules/jest.config.ts
+++ b/tools/eslint-rules/jest.config.ts
@@ -2,6 +2,7 @@
 export default {
   displayName: 'eslint-rules',
   preset: '../../jest.preset.js',
+  testEnvironment: 'node',
   transform: {
     '^.+\\.[tj]s$': ['@swc/jest', {}],
   },