<!--
Thank you for submitting a pull request!

Please verify that:
* [x] Code is up-to-date with the `master` branch
* [ ] Your changes are covered by tests (if possible)
* [x] You've run `yarn change` locally


PR flow tips:
* [ ] Try to start with a Draft PR
* [ ] Once you're ready (ideally the pipeline is passing) promote your PR to Ready for Review. This step will auto-assign reviewers for your PR.
-->

## Previous Behavior

<!-- This is the behavior we have today -->

## New Behavior

Upgraded `@typescript-eslint` to v8 

[Release notes](https://typescript-eslint.io/blog/announcing-typescript-eslint-v8/)


## Related Issue(s)

~~❌  Currently blocked by `nx`. When `nx` will upgrade to the latest eslint and will bump `@typescript-eslint` to v8 as well in `@nx/eslint-plugin` we can proceed forward with this.~~

| Before  | After   |
|--------|--------|
| 11m 5s | 10m 23s / **𝚫 6.3%  🚀  faster** | 


- Partially implements #30334